Add some mouse wheel support.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@10561 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/ewl/src/ewl_callback.h
@@ -49,7 +49,7 @@ struct Ewl_Callback
  * Retrives the callback list from a widget for a certain event type.
  */
 #define EWL_CALLBACK_LIST_POINTER(w, t) \
-		(void *)((unsigned int)(w->callbacks[t]) & \
+		(void *)((long)(w->callbacks[t]) & \
 					~EWL_CALLBACK_NOTIFY_MASK)
 
 /**
@@ -57,15 +57,15 @@ struct Ewl_Callback
  * Retrives the callback flags from a widget for a certain event type.
  */
 #define EWL_CALLBACK_FLAGS(w, t) \
-		((unsigned int)(w->callbacks[t]) & \
+		((long)(w->callbacks[t]) & \
 					EWL_CALLBACK_NOTIFY_MASK)
 /**
  * @def EWL_CALLBACK_FLAG_INTERCEPT(w, t)
  * Sets the callback intercept flag from a widget for a certain event type.
  */
 #define EWL_CALLBACK_FLAG_INTERCEPT(w, t) \
-		((unsigned int)w->callbacks[t] = \
-			 (unsigned int)EWL_CALLBACK_LIST_POINTER(w, t) | \
+		((long)w->callbacks[t] = \
+			 (long)EWL_CALLBACK_LIST_POINTER(w, t) | \
 			 EWL_CALLBACK_NOTIFY_INTERCEPT)
 
 /**
@@ -73,8 +73,8 @@ struct Ewl_Callback
  * Clears the callback intercept flag from a widget for a certain event type.
  */
 #define EWL_CALLBACK_FLAG_NOINTERCEPT(w, t) \
-		((unsigned int)w->callbacks[t] = \
-		 	 (unsigned int)EWL_CALLBACK_LIST_POINTER(w, t) & \
+		((long)w->callbacks[t] = \
+		 	 (long)EWL_CALLBACK_LIST_POINTER(w, t) & \
 			 EWL_CALLBACK_NOTIFY_MASK & ~EWL_CALLBACK_NOTIFY_INTERCEPT)
 
 /**
@@ -82,8 +82,8 @@ struct Ewl_Callback
  * Sets the callback notify flag from a widget for a certain event type.
  */
 #define EWL_CALLBACK_FLAG_NOTIFY(w, t) \
-		((unsigned int)w->callbacks[t] = \
-			 (unsigned int)EWL_CALLBACK_LIST_POINTER(w, t) | \
+		((long)w->callbacks[t] = \
+			 (long)EWL_CALLBACK_LIST_POINTER(w, t) | \
 			 EWL_CALLBACK_NOTIFY_NOTIFY)
 
 /**
@@ -91,8 +91,8 @@ struct Ewl_Callback
  * Sets the callback list for a widget for a certain event type.
  */
 #define EWL_CALLBACK_LIST_ASSIGN(w, t, l) \
-		(unsigned int)w->callbacks[t] = (unsigned int)l | \
-			((unsigned int)w->callbacks[t] & \
+		(long)w->callbacks[t] = (long)l | \
+			((long)w->callbacks[t] & \
 			 EWL_CALLBACK_NOTIFY_MASK)
 
 void            ewl_callbacks_init(void);
